requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
Flask_WTF==1.1.1
WTForms==3.0.1
werkzeug==2.3.7
python-dotenv
wtforms[email]